Added item targeting by name
Should now be able to target items directly by using "${name} or
${shortname}", this can be used on items that have not been picked up as
long as the item is unique, if there are multiple items with the same
hammerID the functionality is reserved for when the items are being held
by someone.

Also added a etransfer entry to the main admin menu and a little command
to print useful information about the items in the map.

Hopefully I didn't forget anything :^(